move consts
